requirements.txt
Flask==2.3.3
yt-dlp==2024.4.9
mutagen==1.47.0
Pillow==10.2.0
requests==2.31.0